refactor(dashboard): extract Court dashboard helpers + template factory (F-B-202 PR-1/10)
Audit ref: imp/audits/2026-05-20/findings/track-b/F-B-202.md

Sprint 2 modularization PR-1 of 10. ``app/dashboard/router.py`` is a
3125-LOC god-object hosting 67 routes across 14 feature areas. This
PR makes room for the per-feature sub-router extraction (PR-2 through
PR-10) by pulling pure helpers + the Jinja2 env factory into two
sibling modules. Mirror of the Mastio dashboard split that landed in
late 2026-Q1 (``mcp_proxy/dashboard/_template_env.py`` + 8 sub-routers).

What moves:

  app/dashboard/_template_env.py (new, 50 LOC):
    build_templates(template_dir, *, extra=None) — every future
    sub-router threads its Jinja2Templates through this factory so
    cross-cutting globals land in one place. Currently a no-op; the
    hook is in place for future plugin gates.

  app/dashboard/_helpers.py (new, 330 LOC):
    - _safe_redirect (audit H-IO-1 open-redirect guard)
    - _ctx (Jinja2 context builder — session, csrf_token, jaeger_url)
    - _SEALED_DETAIL + _is_sealed + _require_sealed_reauth +
      _sealed_mutation_details (audit F-B-2 sealed-org guard)
    - _ID_PATTERN + _validate_id + _validate_webhook_url (validators)
    - _build_audit_event_dict (audit row → template dict)
    - _count_chip (HTMX nav badge)
    - _broker_url_from_request (X-Forwarded-* aware URL builder)
    - _generate_agent_cert (dev cert minter for download-bundle)

  app/dashboard/router.py (3125 → 2930 LOC, -195 LOC):
    - Imports the 13 helpers + build_templates from new modules.
    - Inline duplicates removed; comment breadcrumbs left at the old
      definition sites so a reader following git-blame lands on the
      explanation rather than empty space.
    - templates = build_templates(_TEMPLATE_DIR) replaces the
      direct Jinja2Templates(directory=...) call.

What stays:

  _DISPLAY_NAME_MAX, _CAPABILITY_MAX_LEN, _CAPABILITY_MAX_COUNT, and
  _AUDIT_LIMIT constants stay in router.py — they belong to specific
  endpoint groups (orgs, agents, audit) and will move to their
  per-feature sub-routers in PR-2/5/9 respectively.

Behavioural delta: zero. URLs unchanged, response bodies unchanged,
no route registration change, no template render change. This is
purely a file split. The audit dossier (F-B-202) calls out the
sequence: PR-1 sets up the extraction points, PR-2 through PR-10 do
the actual route moves into per-feature sub-routers.

Tests:
  - pytest tests/test_dashboard*.py -q — 202/203 (1 pre-existing
    deselect: test_dashboard_approvals_nav::test_badge_approvals_empty_when_plugin_unavailable)
  - pytest tests/ --ignore=tests/integration --ignore=tests/e2e
    --deselect <known> -q — 4209 passed, 10 skipped, 32 xfailed
    (2 pre-existing xdist flakes pass on isolated re-run)
  - ./sandbox/demo.sh full + ./sandbox/smoke.sh full — 25/25
    (B4.7 timing flake required one re-run, expected)

Next PR-2 (F-B-202 modularization): extract auth_routes.py
(login/logout/register/OIDC, ~380 LOC, 9 routes).
